refactors to please codeclimate
