+++ src/main/java/ok/dht/test/mikhaylov/MyService.java
@@ -8,7 +8,14 @@ import ok.dht.test.mikhaylov.dao.BaseEntry;
 import ok.dht.test.mikhaylov.dao.Config;
 import ok.dht.test.mikhaylov.dao.Entry;
 import ok.dht.test.mikhaylov.dao.artyomdrozdov.MemorySegmentDao;
-import one.nio.http.*;
+import one.nio.http.HttpServer;
+import one.nio.http.HttpServerConfig;
+import one.nio.http.HttpSession;
+import one.nio.http.Param;
+import one.nio.http.Path;
+import one.nio.http.Request;
+import one.nio.http.RequestMethod;
+import one.nio.http.Response;
 import one.nio.server.AcceptorConfig;
 
 import java.io.IOException;
