+++ src/main/java/ok/dht/test/mikhaylov/dao/artyomdrozdov/Data.java
@@ -1,0 +1,7 @@
+package ok.dht.test.mikhaylov.dao.artyomdrozdov;
+
+import jdk.incubator.foreign.MemorySegment;
+import ok.dht.test.mikhaylov.dao.Entry;
+
+public interface Data extends Iterable<Entry<MemorySegment>> {
+}
